pyproject.toml
[tool.poetry]
name = "cattrs-extras"
version = "0.0.0"
description = "Advanced converters for cattrs"
authors = ["Lev Gorodetskiy <github@droserasprout.space>"]
homepage = "https://github.com/droserasprout/cattrs-extras"
repository = "git@github.com:droserasprout/cattrs-extras.git"
packages = [
    { include = "cattrs_extras", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
cattrs = "^1.0.0"
dateutils = "^0.6.12"
pytimeparse = "^1.1.8"
tortoise-orm = {version = "^0.16.20", optional = true}
typing-extensions = "^3.7.4"
typing-inspect = "^0.6.0"

[tool.poetry.dev-dependencies]
bump2version = "^1.0"
coverage = "^5.1"
isort = {version = "^4.3", extras = ["pyproject"]}
mypy = "0.*"
nose = "^1.3"
nose-timer = "^1.0"
pylint = "^2.6.0"
pylint-exit = "^1.2.0"

[tool.poetry.extras]
tortoise = ["tortoise-orm"]

[tool.nosetests]
verbosity = 2

[tool.isort]
force_single_line = true
line_length = 140

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.masonry.api"
